'Transaction in progress' alert: add external link
